Revert "Remove OoT as being always loaded in ap-worker"
This reverts commit 5b3f6ede8f926c4a41e9c4c3e9240f65a6bd7ca3.

+++ taskcluster/docker/ap-worker/prepare_worlds.sh
@@ -27,6 +27,10 @@ for f in $1/worlds/*; do
         continue
     fi
 
+    # OoT is load bearing for pmd_eos until https://github.com/CrypticMonkey33/ArchipelagoExplorersOfSky/pull/238 gets merged
+    if [[ "$(basename $f)" == "oot" ]]; then
+        continue
+    fi
     (cd $(dirname $f) && zip -r ${DEST}/$(basename $f)-${VERSION}.apworld $(basename $f))
     rm -Rf "$f"
 done
